requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic 
requests
joblib
sqlalchemy 
mysql-connector-python 
xgboost
google-generativeai
python-multipart
